Split up options menu into pages
