+++ src/menu/menu.lua
@@ -9,9 +9,8 @@ local MainMenu = require("src.menu.MainMenu")
 local GameWrapper = require("src.menu.GameWrapper")
 local Highscores = require("src.menu.HighscoresMenu")
 local GameEndScreen = require("src.menu.GameEndScreen")
-local OptionsMenu = require("src.menu.OptionsMenu")
+local OptionsMenu = require("src.menu.options.OptionsMenu")
 local AboutMenu = require("src.menu.AboutMenu")
-local LanguageSelectMenu = require("src.menu.LanguageSelectMenu")
 
 
 local Menu = GUI(
@@ -21,7 +20,6 @@ local Menu = GUI(
         GameEndScreen,
         OptionsMenu,
         AboutMenu,
-        LanguageSelectMenu,
         MainMenu,
         {},
         "switcher"
